Update README.md
In the project, we are using the object detection code using Twilio, Yolo in Python. This is the base code for capturing any object whether it is an animal or human or any non-living object. This detection is possible since we already trained the AI in the code by providing a list of objects and classifications to detect. The Yolo was used as an instant testing and display. You can use the Harcasscade code for the local server